1. qt 6.10.2 works with emscripten 4.0.7 2. no async error

+++ src/clock/mumeclock.cpp
@@ -78,7 +78,11 @@ private:
 
 public:
     NODISCARD static int keyToValue(const QString &key) { return g_qme.keyToValue(key.toUtf8()); }
-    NODISCARD static QString valueToKey(const int value) { return g_qme.valueToKey(value); }
+    NODISCARD static QString valueToKey(const int value)
+    {
+        // Qt 6.10+ changed valueToKey() signature from int to quint64
+        return g_qme.valueToKey(static_cast<quint64>(value));
+    }
     NODISCARD static int keyCount() { return g_qme.keyCount(); }
 };
